[PC-894] refactor: dto 디렉터리 위치 통일
