+++ admin/src/main/java/org/yapp/match/presentation/request/ManualMatchCancelRequest.java
@@ -1,4 +1,4 @@
-package org.yapp.match.dto.request;
+package org.yapp.match.presentation.request;
 
 public record ManualMatchCancelRequest(
     Long manualMatchId
