+++ admin/src/main/java/org/yapp/match/presentation/response/ManualMatchCandidateResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.match.dto.response;
+package org.yapp.match.presentation.response;
 
 public record ManualMatchCandidateResponse(Long userId, String nickName, Boolean isMatched) {
